+++ lib/api/entrata/update-wp-floorplan-entrata.php
@@ -36,6 +36,7 @@ function rfs_entrata_update_floorplan_meta( $args, $floorplan_data, $units ) {
 		);
 		
 		$success = update_post_meta( $args['wordpress_floorplan_post_id'], 'api_response', $api_response );
+		rfs_mark_sync_failed( $args['wordpress_floorplan_post_id'], 'floorplans_api' );
 		return;
 	}
 
@@ -56,6 +57,7 @@ function rfs_entrata_update_floorplan_meta( $args, $floorplan_data, $units ) {
 		);
 
 		$success = update_post_meta( $args['wordpress_floorplan_post_id'], 'api_response', $api_response );
+		rfs_mark_sync_failed( $args['wordpress_floorplan_post_id'], 'floorplans_api' );
 
 		return;
 	}
@@ -170,6 +172,8 @@ function rfs_entrata_update_floorplan_meta( $args, $floorplan_data, $units ) {
 	foreach ( $meta as $key => $value ) {
 		$success = update_post_meta( $args['wordpress_floorplan_post_id'], $key, $value );
 	}
+
+	rfs_mark_sync_succeeded( $args['wordpress_floorplan_post_id'], 'floorplans_api' );
 }